+++ docs/conf.py
@@ -54,7 +54,7 @@ copyright = u'2015, Barney Gale'
 # built documents.
 #
 # The short X.Y version.
-version = '0.8'
+version = '0.9'
 # The full version, including alpha/beta/rc tags.
 release = version