+++ CHANGES.rst
@@ -5,6 +5,10 @@ master
 ------
 
 - Added support for Minecraft 1.12.2.
+- Added documentation for proxies
+- Added a "fast forwarding" mode for proxies that skips packing/unpacking of
+  packets.
+- Re-arranged some proxy internals.
 - Replaced ``quarry.net.tasks`` with ``quarry.net.ticker``. An instance of the
   ``Ticker`` class is available as ``self.ticker`` from protocols. This object
   has ``add_delay()`` and ``add_loop()`` methods for setting up delayed and